+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## V4.0.2
 
 - Adds `info` prop to `<InfoListItem>` to support a third line of text.
+- Change title and description types from `string` to `ReactNode`.
 
 ## v4.0.1